improve: use new GitHub based binary repo for KubeAPITest
