+++ junit/kube-api-test/core/src/main/java/io/fabric8/kubeapitest/binary/BinaryManager.java
@@ -40,7 +40,7 @@ public class BinaryManager {
   public BinaryManager(KubeAPIServerConfig config) {
     this.config = config;
     this.osInfo = new OSInfo();
-    this.downloader = new BinaryDownloader(config.getKubeAPITestDir(), osInfo);
+    this.downloader = new CloudStorageBinaryBinaryDownloader(config.getKubeAPITestDir(), osInfo);
   }
 
   public void initAndDownloadIfRequired() {
