+++ junit/kube-api-test/core/src/main/java/io/fabric8/kubeapitest/binary/GitHubBinaryDownloader.java
@@ -1,0 +1,34 @@
+package io.fabric8.kubeapitest.binary;
+
+import java.io.File;
+
+public class GitHubBinaryDownloader implements BinaryDownloader {
+
+
+  @Override
+  public File download(String version) {
+
+
+    return null;
+  }
+
+  @Override
+  public File downloadLatest() {
+    return null;
+  }
+
+  @Override
+  public File downloadLatestWildcard(String wildcardVersion) {
+    return null;
+  }
+
+  @Override
+  public String findLatestVersion() {
+    return "";
+  }
+
+  @Override
+  public String findLatestOfWildcard(String wildcardVersion) {
+    return "";
+  }
+}
